add get all requests for shoppingLists and Items, id not working yet

+++ backend/src/main/java/org/example/backend/Model/ShoppingListEntry.java
@@ -1,7 +1,8 @@
 package org.example.backend.Model;
 
 public record ShoppingListEntry(
-        Item item,
-        double quantity
-) {
-}
+    Item item,  // Nested Item object
+    int quantity
+) { }
+
+
